Planning: Split out into multiple files
- Move evolutionary Algorithm classes to own file
- Move the remaining Problem classes to another name
